requirements.txt
aiohttp==3.9.3
beautifulsoup4==4.12.3
crewai==0.11.0
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langchain-community
langchain-openai==0.0.2.post1
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0
tiktoken==0.5.2
unstructured
streamlit

aiofiles==24.1.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
crewai==0.11.0
cryptography==43.0.3
dataclasses-json==0.6.7
Deprecated==1.2.15
distro==1.9.0
docstring-parser==0.15
emoji==2.14.0
eval_type_backport==0.2.0
filetype==1.2.0
frozenlist==1.5.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
instructor==0.5.2
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.2.post1
langdetect==1.0.9
langsmith==0.0.87
lxml==5.3.0
markdown-it-py==3.0.0
marshmallow==3.23.1
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==1.12.0
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
packaging==23.2
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pypdf==5.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2024.10.22
python-magic==0.4.27
python-oxmsg==0.0.1
PyYAML==6.0.2
RapidFuzz==3.10.1
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.9.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
tenacity==8.5.0
tiktoken==0.5.2
tqdm==4.67.1
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
unstructured==0.16.6
unstructured-client==0.28.0
urllib3==2.2.3
webencodings==0.5.1
wrapt==1.17.0
yarl==1.18.0
zipp==3.21.